Remove WithProvisioningFolderMetadata from shared server
Moved CreateFolder_FolderMetadataFlag and FolderMetadataFileProtection
tests to the foldermetadata package where they belong. These tests
require the feature flag and now use their own RunGrafana(t) calls.

Restored the original ProvisionedFolders test assertions (403 Forbidden)
now that the shared server no longer has the flag enabled.

Made-with: Cursor
