+++ pkg/tests/apis/provisioning/files_test.go
@@ -1088,215 +1088,4 @@ func TestIntegrationProvisioning_FilesAuthorization(t *testing.T) {
 	})
 }
 
-func TestIntegrationProvisioning_CreateFolder_FolderMetadataFlag(t *testing.T) {
-	helper := sharedHelper(t)
-	ctx := context.Background()
-
-	const repo = "folder-metadata-test-repo"
-	helper.CreateRepo(t, common.TestRepo{Name: repo, Target: "instance", SkipResourceAssertions: true})
-
-	addr := helper.GetEnv().Server.HTTPServer.Listener.Addr().String()
-
-	postFolder := func(t *testing.T, path string) *http.Response {
-		t.Helper()
-		u := fmt.Sprintf("http://admin:admin@%s/apis/provisioning.grafana.app/v0alpha1/namespaces/default/repositories/%s/files/%s", addr, repo, path)
-		req, err := http.NewRequest(http.MethodPost, u, nil)
-		require.NoError(t, err)
-		resp, err := http.DefaultClient.Do(req)
-		require.NoError(t, err)
-		return resp
-	}
-
-	t.Run("simple folder creation writes _folder.json with stable UID", func(t *testing.T) {
-		resp := postFolder(t, "meta-test-folder/")
-		// nolint:errcheck
-		defer resp.Body.Close()
-		require.Equal(t, http.StatusOK, resp.StatusCode, "creating folder should succeed")
-
-		wrapObj, err := helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "meta-test-folder/_folder.json")
-		require.NoError(t, err, "_folder.json should be readable via the files endpoint")
-
-		apiVersion, _, _ := unstructured.NestedString(wrapObj.Object, "resource", "file", "apiVersion")
-		require.Equal(t, "folder.grafana.app/v1beta1", apiVersion)
-		kind, _, _ := unstructured.NestedString(wrapObj.Object, "resource", "file", "kind")
-		require.Equal(t, "Folder", kind)
-		folderUID, _, _ := unstructured.NestedString(wrapObj.Object, "resource", "file", "metadata", "name")
-		require.NotEmpty(t, folderUID, "_folder.json should contain a non-empty stable UID")
-		title, _, _ := unstructured.NestedString(wrapObj.Object, "resource", "file", "spec", "title")
-		require.Equal(t, "meta-test-folder", title)
-
-		_, err = helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "meta-test-folder/.keep")
-		require.Error(t, err, ".keep should not exist when flag is enabled")
-
-		_, err = helper.Folders.Resource.Get(ctx, folderUID, metav1.GetOptions{})
-		require.NoError(t, err, "Grafana folder should exist with the stable UID from _folder.json")
-	})
-
-	t.Run("nested creation writes _folder.json for every folder in the path", func(t *testing.T) {
-		resp := postFolder(t, "parent-folder/child-folder/")
-		// nolint:errcheck
-		defer resp.Body.Close()
-		require.Equal(t, http.StatusOK, resp.StatusCode, "creating nested folder should succeed")
-
-		// Parent must have _folder.json
-		parentWrap, err := helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "parent-folder/_folder.json")
-		require.NoError(t, err, "parent _folder.json should exist")
-		parentUID, _, _ := unstructured.NestedString(parentWrap.Object, "resource", "file", "metadata", "name")
-		require.NotEmpty(t, parentUID)
-
-		// Child must have _folder.json
-		childWrap, err := helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "parent-folder/child-folder/_folder.json")
-		require.NoError(t, err, "child _folder.json should exist")
-		childUID, _, _ := unstructured.NestedString(childWrap.Object, "resource", "file", "metadata", "name")
-		require.NotEmpty(t, childUID)
-		require.NotEqual(t, parentUID, childUID, "each folder gets a distinct UID")
-
-		childAPIVersion, _, _ := unstructured.NestedString(childWrap.Object, "resource", "file", "apiVersion")
-		require.Equal(t, "folder.grafana.app/v1beta1", childAPIVersion)
-		childTitle, _, _ := unstructured.NestedString(childWrap.Object, "resource", "file", "spec", "title")
-		require.Equal(t, "child-folder", childTitle)
-
-		_, err = helper.Folders.Resource.Get(ctx, childUID, metav1.GetOptions{})
-		require.NoError(t, err, "child Grafana folder should exist with the stable UID")
-	})
-
-	t.Run("duplicate folder creation returns 409 Conflict", func(t *testing.T) {
-		resp := postFolder(t, "duplicate-folder/")
-		// nolint:errcheck
-		defer resp.Body.Close()
-		require.Equal(t, http.StatusOK, resp.StatusCode, "first creation should succeed")
-
-		resp2 := postFolder(t, "duplicate-folder/")
-		// nolint:errcheck
-		defer resp2.Body.Close()
-		require.Equal(t, http.StatusConflict, resp2.StatusCode, "second creation should return 409 Conflict")
-	})
-
-	t.Run("child created inside existing managed folder gets its own _folder.json", func(t *testing.T) {
-		resp := postFolder(t, "managed-parent/")
-		// nolint:errcheck
-		defer resp.Body.Close()
-		require.Equal(t, http.StatusOK, resp.StatusCode, "creating parent folder should succeed")
-
-		parentWrap, err := helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "managed-parent/_folder.json")
-		require.NoError(t, err, "parent _folder.json should exist")
-		parentUID, _, _ := unstructured.NestedString(parentWrap.Object, "resource", "file", "metadata", "name")
-		require.NotEmpty(t, parentUID, "parent should have a non-empty stable UID")
-
-		resp2 := postFolder(t, "managed-parent/child-folder/")
-		// nolint:errcheck
-		defer resp2.Body.Close()
-		require.Equal(t, http.StatusOK, resp2.StatusCode, "creating child folder should succeed")
-
-		childWrap, err := helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "managed-parent/child-folder/_folder.json")
-		require.NoError(t, err, "child _folder.json should exist")
-		childUID, _, _ := unstructured.NestedString(childWrap.Object, "resource", "file", "metadata", "name")
-		require.NotEmpty(t, childUID, "child should have a non-empty stable UID")
-		require.NotEqual(t, parentUID, childUID, "child and parent UIDs must differ")
-
-		// Parent _folder.json must be unchanged.
-		parentWrap2, err := helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "managed-parent/_folder.json")
-		require.NoError(t, err, "parent _folder.json should still exist after child creation")
-		parentUID2, _, _ := unstructured.NestedString(parentWrap2.Object, "resource", "file", "metadata", "name")
-		require.Equal(t, parentUID, parentUID2, "parent UID must be unchanged after child creation")
-
-		_, err = helper.Folders.Resource.Get(ctx, parentUID, metav1.GetOptions{})
-		require.NoError(t, err, "parent Grafana folder should exist")
-		_, err = helper.Folders.Resource.Get(ctx, childUID, metav1.GetOptions{})
-		require.NoError(t, err, "child Grafana folder should exist")
-	})
-}
-
-func TestIntegrationProvisioning_FolderMetadataFileProtection(t *testing.T) {
-	helper := sharedHelper(t)
-	ctx := context.Background()
-
-	const repo = "folder-protection-test-repo"
-	helper.CreateRepo(t, common.TestRepo{Name: repo, Target: "instance", SkipResourceAssertions: true})
-
-	addr := helper.GetEnv().Server.HTTPServer.Listener.Addr().String()
-	filesURL := func(filePath string) string {
-		return fmt.Sprintf("http://admin:admin@%s/apis/provisioning.grafana.app/v0alpha1/namespaces/default/repositories/%s/files/%s",
-			addr, repo, filePath)
-	}
-
-	// Create a managed folder so its _folder.json exists for PUT/DELETE tests.
-	req, err := http.NewRequest(http.MethodPost, filesURL("protected-folder/"), nil)
-	require.NoError(t, err)
-	resp, err := http.DefaultClient.Do(req)
-	require.NoError(t, err)
-	// nolint:errcheck
-	defer resp.Body.Close()
-	require.Equal(t, http.StatusOK, resp.StatusCode, "setup: creating protected-folder should succeed")
-
-	t.Run("POST to _folder.json is blocked", func(t *testing.T) {
-		body := []byte(`{"apiVersion":"folder.grafana.app/v1beta1","kind":"Folder","metadata":{"name":"some-uid"},"spec":{"title":"attempt"}}`)
-		req, err := http.NewRequest(http.MethodPost, filesURL("new-folder/_folder.json"), bytes.NewReader(body))
-		require.NoError(t, err)
-		req.Header.Set("Content-Type", "application/json")
-		resp, err := http.DefaultClient.Do(req)
-		require.NoError(t, err)
-		// nolint:errcheck
-		defer resp.Body.Close()
-		require.Equal(t, http.StatusForbidden, resp.StatusCode, "direct POST to _folder.json must be blocked")
-	})
-
-	t.Run("PUT to existing _folder.json is blocked", func(t *testing.T) {
-		body := []byte(`{"apiVersion":"folder.grafana.app/v1beta1","kind":"Folder","metadata":{"name":"tampered-uid"},"spec":{"title":"tampered"}}`)
-		req, err := http.NewRequest(http.MethodPut, filesURL("protected-folder/_folder.json"), bytes.NewReader(body))
-		require.NoError(t, err)
-		req.Header.Set("Content-Type", "application/json")
-		resp, err := http.DefaultClient.Do(req)
-		require.NoError(t, err)
-		// nolint:errcheck
-		defer resp.Body.Close()
-		require.Equal(t, http.StatusForbidden, resp.StatusCode, "PUT to _folder.json must be blocked")
-	})
-
-	t.Run("DELETE of _folder.json is blocked", func(t *testing.T) {
-		req, err := http.NewRequest(http.MethodDelete, filesURL("protected-folder/_folder.json"), nil)
-		require.NoError(t, err)
-		resp, err := http.DefaultClient.Do(req)
-		require.NoError(t, err)
-		// nolint:errcheck
-		defer resp.Body.Close()
-		require.Equal(t, http.StatusForbidden, resp.StatusCode, "DELETE of _folder.json must be blocked")
-	})
-
-	t.Run("GET of _folder.json is still allowed", func(t *testing.T) {
-		wrapObj, err := helper.Repositories.Resource.Get(ctx, repo, metav1.GetOptions{}, "files", "protected-folder/_folder.json")
-		require.NoError(t, err, "_folder.json must remain readable")
-		uid, _, _ := unstructured.NestedString(wrapObj.Object, "resource", "file", "metadata", "name")
-		require.NotEmpty(t, uid)
-	})
-
-	rootFolderBody := []byte(`{"apiVersion":"folder.grafana.app/v1beta1","kind":"Folder","metadata":{"name":"root-uid"},"spec":{"title":"root"}}`)
-
-	for _, tc := range []struct {
-		name   string
-		method string
-		body   []byte
-	}{
-		{name: "POST to root _folder.json is blocked", method: http.MethodPost, body: rootFolderBody},
-		{name: "PUT to root _folder.json is blocked", method: http.MethodPut, body: rootFolderBody},
-		{name: "DELETE of root _folder.json is blocked", method: http.MethodDelete},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			var bodyReader io.Reader
-			if tc.body != nil {
-				bodyReader = bytes.NewReader(tc.body)
-			}
-			req, err := http.NewRequest(tc.method, filesURL("_folder.json"), bodyReader)
-			require.NoError(t, err)
-			if tc.body != nil {
-				req.Header.Set("Content-Type", "application/json")
-			}
-			resp, err := http.DefaultClient.Do(req)
-			require.NoError(t, err)
-			// nolint:errcheck
-			defer resp.Body.Close()
-			require.Equal(t, http.StatusForbidden, resp.StatusCode)
-		})
-	}
-}
